+++ pages/labelingTool.py
@@ -1262,7 +1262,7 @@ class MainWindow(QtWidgets.QMainWindow, WindowMixin):
 		self.set_scroll(orientation, value)
 
 	def set_scroll(self, orientation, value):
-		self.scroll_bars[orientation].setValue(round(value))
+		self.scroll_bars[orientation].setValue(np.int32(round(value)))
 
 	def set_zoom(self, value):
 		self.actions.fitWidth.setChecked(False)
